Implemented control.cmd and control.event handles
Usage of the control.* special handles is now detected during
svc_set_fds, and then at service-start time I check those flags
and allocate a controller if needed, and the pipes for it.
On failure, the pipes and controller are deleted and the service
start is attempted again in FORK_RETRY_DELAY seconds.

Added a simple test case that the pipes attached to a controller
service work.

fixes #8, fixes #9