Rough draft of config-sync.pl, to apply a yaml config to running daemonproxy
This script reads a yaml config file and then interacts with daemonproxy
on stdin/stdout to make daemonproxy's state match the config.  If the config
file has not changed from the previous run, it should not alter daemonproxy's
state in any way (except to decrement the SIGHUP count)

This code is incomplete, and lightly commented.

refs #45, refs #47, refs #49