Eliminated START_PENDING from service state machine
On a service start, the service state machine used to transition
to SERVICE_STATE_START_PENDING and then to SERVICE_STATE_START,
but there really wasn't much point in that because both states
looked the same in the event, and used the same logic to decide
when to restart.

This simplifies the code a bit.