Added new control.socket special FD
The pipes between daemonproxy and its controller are now
actually sockets.  This is in preparation for a new command
that passes fds as ancilliary data over a unix socket.

If the service uses control.socket, it gets a bi-directional
socket, but if it uses only one of control.event or
control.cmd then the un-used direction will be shutdown()
before forking.