Assert that testcases never have newline or tab in send() arguments
Originally, send was a free form string.  Then I made it so that
individual arguments get joined together as a single protocol line.
Now I'm also ensuring that no old code was left over which
accidentally sent newlines in the middle of the command.

This was particularly important for the pipe test, because I expanded
the embedded perl script to multiple lines but the newline ended the
command, so it became "perl -e ''", which always returns 0.

Also added a ->terminate_ok method to handle the common case of

    $dp->send( 'terminate', 0 );
    $dp->exit_is( 0 );