New strseg_split_1 to replace strseg_tok_next
The strseg_tok_next has a screwy API, and I'm trying to
see if I can replace it with something more sensible.

To start with, I'll use strseg_split_1 which is nearly
the same except the first argument gets the prefix and
the remainder is stored in the second arg.