Fixed handling of empty config file lines; added testcase
Previously code would treat lines starting with whitespace as commants,
and warn that empty lines were invalid commands.  It now ignores the
empty lines and only ignores a whitespace line if the entire line is
whitespace.

Added a testcase that runs daemonproxy with a configfile to test these
cases and check that the correct errors are generated.

fixes #51