Print usage on interactive start, and log running services on exit
Daemonproxy now prints a version and usage blurb at the start of
interactive mode, to help orient anyone not familiar with it.

When exiting interactive mode, it logs a message that it is (or
isn't) terminating, and logs each service which is still running.

Also updated several test cases which didn't like the extra lines
of input at startup.

fixes #50